+++ scripts/lichessTools-init.js
@@ -147,6 +147,7 @@ lichessTools.loadTool(LiChessTools.Tools.TeamStats);
 lichessTools.loadTool(LiChessTools.Tools.CalculationTrainer);
 lichessTools.loadTool(LiChessTools.Tools.ExplorerChessagine);
 lichessTools.loadTool(LiChessTools.Tools.ShowFps);
+lichessTools.loadTool(LiChessTools.Tools.ShowOpponentMoves);
 
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);