+++ HISTORY.md
@@ -1,5 +1,9 @@
 History of features added to LiChess Tools in time (reversed order)
 
+01 Jun
+
+- Show opponent legal moves
+
 29 May
 
 - Study chapter coverage